update: エラーハンドリングの強化
