+++ lambda_functions/retrieval/core/main/orchestration/lambda_orchestration.py
@@ -1,14 +1,8 @@
-import json
 import logging
-from typing import Optional, Union, List, Any
+from typing import Union, List, Any
 from uuid import UUID
-from pydantic import BaseModel, ValidationError
 
-from core.base import (
-    GenerationConfig,
-    GenericMessageResponse,
-    R2RException
-)
+from lambda_functions.common.core.main.exception import LambdaException
 from core.base.api.models import UserResponse
 from core.main.services.auth_service import AuthService
 from core.main.services.retrieval_service import RetrievalService
@@ -17,26 +11,13 @@ from shared.abstractions.search import VectorSearchSettings, KGSearchSettings
 logger = logging.getLogger()
 
 
-class LambdaException(Exception):
-    def __init__(self, status_code: int, error_msg: str):
-        self.status_code = status_code
-        self.error_msg = error_msg
-
-    def __str__(self):
-        obj = {
-            "statusCode": self.status_code,
-            "errorMessage": self.error_msg
-        }
-        return json.dumps(obj)
-
-
 def handle_error(func):
     def inner(*args, **kwargs):
         try:
             return func(*args, **kwargs)
         except Exception as err:
             logger.error(str(err))
-            raise LambdaException(500, err["errorMessage"])
+            raise LambdaException(err["errorMessage"], 500)
     return inner
 
 
@@ -52,6 +33,7 @@ class LambdaOrchestration:
     ):
         self.services: Services = services
 
+    @handle_error
     def _select_filters(
         self,
         auth_user: UserResponse,
